requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.0.0
mysqlx-connector-python==9.2.0
packaging==24.2
pluggy==1.5.0
protobuf==4.25.3
pytest==8.3.5
python-dotenv==1.0.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
